pyproject.toml
[project]
name = "weiboanalysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "jieba>=0.42.1",
    "matplotlib>=3.10.8",
    "mysqlclient>=2.2.8",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "snownlp>=0.12.3",
    "sqlalchemy>=2.0.47",
    "wordcloud>=1.9.6",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
